# contracts/openapi/nex-oa.openapi.yaml
openapi: 3.1.0
info:
  title: NeX OA API
  version: 0.0.0-slice1269
...
paths:
...
  /internal/v1/auth/service-claim:
    get:
      operationId: validateOaServiceClaim
      tags:
        - Platform
      responses:
        "200":
          description: Service claim is valid for this service.
        "401":
          description: Service claim is missing or invalid.
  /api/v1/auth/service-token:
    post:
      operationId: exchangeOaClientCredentialToken
      tags: [Signed Tokens]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceTokenExchangeRequest"
      responses:
        "200":
          description: Five-minute RS256 service access token issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignedTokenResponse"
        "400":
          description: Service token request is invalid.
        "401":
          description: Client credential is invalid or inactive.
        "403":
          description: Requested audience or scope is not allowed.
        "503":
          description: Signing key or external custody is unavailable.
  /api/v1/auth/introspect:
    post:
      operationId: introspectOaSignedServiceToken
      tags: [Signed Tokens]
      security:
        - signedServiceBearer: []
      x-nex-required-scopes: [token:introspect]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenIntrospectionRequest"
      responses:
        "200":
          description: Privacy-bounded signed-token introspection result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenIntrospectionResponse"
        "400": {description: Introspection request is invalid.}
        "401": {description: Caller token is missing or invalid.}
        "403": {description: Caller lacks token introspection scope.}
...
components:
...
  securitySchemes:
    serviceBearer:
      type: http
      scheme: bearer
      bearerFormat: NeX service claim
    signedServiceBearer:
      type: http
      scheme: bearer
      bearerFormat: RS256 at+jwt
  schemas:
    OaServiceId:
      type: string
      enum: [nex-oa, nex-ag, nex-ae-api, nex-cx, nex-mo]
    ServiceTokenExchangeRequest:
      type: object
      additionalProperties: false
      required: [grant_type, credential_id, client_secret, audience, scope]
      properties:
        grant_type: {const: client_credentials}
        credential_id: {type: string, pattern: "^[a-z][a-z0-9._-]{1,63}$"}
        client_secret: {type: string, minLength: 1, maxLength: 256, writeOnly: true}
        audience: {$ref: "#/components/schemas/OaServiceId"}
        scope: {type: string, minLength: 3, maxLength: 1024}
    TokenIntrospectionRequest:
      type: object
      additionalProperties: false
      required: [token, audience]
      properties:
        token: {type: string, minLength: 64, maxLength: 16384, writeOnly: true}
        audience: {$ref: "#/components/schemas/OaServiceId"}
        required_scopes:
          type: array
          maxItems: 128
          uniqueItems: true
          items: {type: string, pattern: "^[a-z][a-z0-9._-]*:[a-z][a-z0-9._-]*$"}
...
    SignedTokenResponse:
      type: object
      x-nex-canonical-json-schema: schemas/service/nex_oa/signed_token_response.v1.schema.json
    TokenIntrospectionResponse:
      type: object
      x-nex-canonical-json-schema: schemas/service/nex_oa/token_introspection.v1.schema.json
...